Fixing bson mapping

+++ src/internal/entity/dto/place.go
@@ -2,6 +2,6 @@ package dto
 
 type PlaceDtl struct {
 	ID        int64   `json:"id"`
-	Latitude  float32 `json:"lat"`
-	Longitude float32 `json:"long"`
+	Latitude  float64 `json:"lat"`
+	Longitude float64 `json:"long"`
 }
